requirements.txt
Flask==2.2.5
pdf2docx==0.5.6
pdf2image==1.16.3
Pillow==10.1.0
Pillow==10.1.0
reportlab==4.0.8
